Renamed and moved BugTrackerEventTrigger into a proper package.
